+++ examples/react/vite.config.ts
@@ -18,15 +18,15 @@ export default defineConfig({
   resolve: {
     alias: [
       {
-        find: "@moritzbrantner/editor-core/react",
+        find: "@moenarch/editor-core/react",
         replacement: resolve(rootDir, "src/react.tsx"),
       },
       {
-        find: "@moritzbrantner/editor-core/tree",
+        find: "@moenarch/editor-core/tree",
         replacement: resolve(rootDir, "src/tree.ts"),
       },
       {
-        find: "@moritzbrantner/editor-core",
+        find: "@moenarch/editor-core",
         replacement: resolve(rootDir, "src/index.ts"),
       },
     ],